Renaming files to make internal
